# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreateWorkDTO:
    properties:
      company:
        maxLength: 100
        minLength: 2
        type: string
      employee_id:
        type: string
      end_date:
        type: string
      position:
        maxLength: 100
        minLength: 2
        type: string
      responsibilities:
        maxLength: 500
        type: string
      start_date:
        type: string
    required:
    - company
    - employee_id
    - position
    - start_date
    type: object
...
  dtos.UpdateWorkDTO:
    properties:
      company:
        maxLength: 100
        minLength: 2
        type: string
      end_date:
        type: string
      position:
        maxLength: 100
        minLength: 2
        type: string
      responsibilities:
        maxLength: 500
        type: string
      start_date:
        type: string
    type: object
  dtos.WorkResponseDTO:
    properties:
      company:
        type: string
      created_at:
        type: string
      employee_id:
        type: string
      end_date:
        type: string
      id:
        type: string
      position:
        type: string
      responsibilities:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.FieldError:
    properties:
      field:
        type: string
      label:
        type: string
      message:
        type: string
      tag:
        type: string
      value: {}
    type: object
  utils.Payload:
    properties:
      error:
        description: 'ex: validation_failed, conflict, internal'
        type: string
      fields:
        description: detalhes por campo
        items:
          $ref: '#/definitions/utils.FieldError'
        type: array
      message:
        description: resumo humanizado
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: suporte@empresa.com
    name: Equipa de Desenvolvimento
  description: API de Gestão de Recursos Humanos (Províncias, Municípios, Departamentos,
    etc.)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost:8080/terms
  title: HRMS API
  version: "1.0"
paths:
...
  /employees/{employee_id}/work_history:
    get:
      parameters:
      - description: ID do Funcionário
        in: path
        name: employee_id
        required: true
        type: string
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.WorkResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Listar históricos profissionais de um funcionário
      tags:
      - Histórico Profissional
...
  /work_history:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do histórico profissional
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateWorkDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.WorkResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Criar novo histórico profissional
      tags:
      - Histórico Profissional
  /work_history/{id}:
    delete:
      parameters:
      - description: ID do histórico
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Remover histórico profissional
      tags:
      - Histórico Profissional
    get:
      parameters:
      - description: ID do histórico
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.WorkResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Buscar histórico profissional por ID
      tags:
      - Histórico Profissional
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do histórico
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateWorkDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.WorkResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Payload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Payload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Payload'
      summary: Atualizar histórico profissional
      tags:
      - Histórico Profissional
schemes:
- http
swagger: "2.0"